Add a test that CMake tool variables are set.
This was reported broken but that does not appear to be the case. Adding
the regression test anyway as proof. It might be the case that this is
broken with older CMake versions or perhaps in older NDK releases.

Bug: https://github.com/android/ndk/issues/1698
Test: This is the test
Change-Id: Ia2c69b1263187639d6cf69dcbdb3345f6ee3be45
(cherry picked from commit 5fdfb533b3b75bb0cf3aa910f370218cb9cc3d28)
Merged-In: Ia2c69b1263187639d6cf69dcbdb3345f6ee3be45