Hack for Nexus 7 (2013) not correctly reporting idiv.
2013 Nexus 7s have a Qualcomm Krait processor that, like the Nexus 4,
doesn't correctly report the presence of the idiv instruction.

In general, we would prefer that people not use the deprecated,
unsupported cpufeatures included with the NDK. But we do have a test for
it that is failing, and this seems like the simplest way to get it to pass,
and preferable to just deleting the test.

See https://github.com/android/ndk/issues/1605

Change-Id: I933fb55e001eeee120289e37239b4027b6bfe1e3