Test on API 21 devices.
To prepare for dropping kitkat support, make sure our tests are all
behaving on API 21.

With the exception of ASan, all the changes here were to fix test bugs,
not actual bugs. ASan still needs to be investigated, but I suspect that
one is also a test bug (missing LD_PRELOAD).

Bug: https://github.com/android/ndk/issues/1751
Test: treehugger
Change-Id: I59462c8d046b4aac17fd36f5ab1831867ff02cbe