Yell louder when all devices are missing.
0/0 tests pass is easy to miss. This was historically a "passing" result
to accommodate the build servers where it was expected that no tests
could run because we can't attach devices to the build server. That all
runs through checkbuild.py now, so we can add this check to
run_tests.py.

It's worth noting that this is still not going to help you every time.
We do not expect everyone to have every tested device variant connected
for every test run (it's extremely difficult to do so with an M1 Mac;
the x86/x86_64 emulators aren't available and devices are very hard to
come by), so we don't consider it a failure if only a subset of the
expected devices were connected. If you want to be sure a specific
device configuration was checked when testing, pay attention to the
warnings that are printed during device discovery.

Bug: None
Test: ./run_tests.py with devices unplugged, and with devices connected
Change-Id: Ib794499c4331039c941c9129e8eb6a041c0a3467