Bundle all tests for a device into a single zip.
For now, we produce extra Tradefed zip files along with the old way of
doing things.

The reason for doing this is to reduce the number of test devices we
need to use and potentially flash. This should speed up the tests and
reduce wear on the devices.

Tested by running the corresponding pending CI config changes against a presubmit build of this code.

Change-Id: I2ed46f8f544f507eddc4c4f34862e5801e749cef