Rerun failing tests to capture logcat.
We can't do this on the initial run because we run multiple tests on
each connected device, so there's no (simple) way to associate a portion
of the log with each simultaneous test.

Instead, since there should be a small number of failures, rerun any
failed tests in series (per device; sharding will still work) so we can
be sure the log won't be affected by other tests. It might still be full
of spam from whatever the OS is up to, but we can clear the log before
re-running the test to minimize that.

Bug: None
Test: added an abort to the math test, saw the stack trace
Change-Id: I026e4cfea2af1d0f659309f17dc7c2c0209f4ce8