Produce JSON file with information on tests.
tests.json is a build artifact with test names and commands, and whether
they are broken or unsupported.

The test binaries themselves remain in ndk-tests.tar.bz2, which is already a
build artifact.

The structure of tests.json is a mapping from test config (e.g.
armeabi-v7a-24-new) to a list of tests for that config.

Each test is represented as:

{
  "name": <test name>,
  "cmd": <command to run with 'adb shell'>,
  "broken": [
    <list of device API levels for which the test is broken>
  ],
  "unsupported": [
    <list of device API levels for which the test is unsupported>
  ],
}

Change-Id: I37521051005af262b89d54d779db4095af90ece7