Fix build/runtime API level for static exe tests.
Static executable tests need this to run correctly. They need to build
against the newest API level (there's only one libc.a, which the headers
and CRT objects need to match), but can run against any API level.

We've been getting away with this for a while because it wasn't as
strict, but libc.a now has a dependency on ELF TLS, so we need the newer
CRT objects. We were further getting away with it because --gc-sections
was stripping the ELF TLS dependency.

Bug: https://github.com/android/ndk/issues/1813
Test: this is the test, now it passes
Change-Id: Icc984f9fa3dc83ecd8c3dca8789be55bcb487119