Do not pass API level explicitly for tests.
I'm not sure we've ever needed to do this, and doing so breaks the new
behavior introduced by `override_runtime_minsdkversion` in
https://android-review.git.corp.google.com/c/platform/ndk/+/2392414, as
the override version would be passed and override the value set in
Application.mk.

Tests missed this because it didn't affect CMake (CMake has to override
the default with `extra_cmake_flags`, which take precedence over this
anyway), and the tests that use this won't otherwise fail for ndk-build
(building for API 21 doesn't cause any issues for release builds because
`--gc-sections` is protecting those from the TLS sections, and ndk-build
is release by default).

I haven't added a test for this because what it actually needs for a
useful test is a Python test (it's awkward to do in a regular NDK test),
but those are not in CI, and as such are currently broken anyway. It
would likely be heavily mocked or require some refactoring of
`_get_or_infer_app_platform` anyway. It's maybe worth doing, but we need
to fix the bitrot first.

Bug: None
Test: Manual inspection (see rationale above)
Change-Id: If50f1a845e25311d8fe872284bae3f9ea20eba6d