Update LLVM to clang-r487747b.
Includes a bunch of test `#include` fixes for the reason explained in
the changelog.

Also updates the ASan test. The ASan failure message has been rephrased
so the old test no longer matched the death message correctly. As best
as I can tell the test we have is a copy of the upstream test with all
the supporting functions from the various sanitizer test support headers
copied into it. I've replaced the guts of the test and copied the new
required supporting functions.

This also disables the TSan smoke test. It hangs indefinitely, and we
don't have timeout support (`timeout` is not available on old devices).
The test was already an xfail, so we can just disable it for now. The
test will still be built, so we can be sure that it won't suffer build-
time regressions.

A few libc++ tests were disabled because the tests (either the failing
asserts or whole tests) were removed upstream.

Bug: https://github.com/android/ndk/issues/1298
Bug: https://github.com/android/ndk/issues/1677
Bug: https://github.com/android/ndk/issues/1834
Bug: https://github.com/android/ndk/issues/1860
Bug: https://github.com/android/ndk/issues/1862
Test: checkbuild.py && run_tests.py
Change-Id: Id467537487908cad99099bfeb2cfacf4787d875c