Replace some Arch uses with Abi.
Architecture is a pretty useless abstraction for us. Replacing the
Platforms.get_arches uses with ABI will help with another cleanup I'm
working on. This also moves the logic for determining what ABIs are
available per-API level to ndk.abis, since it really has nothing to do
with the (now dead and gone) platforms directory.

Bug: None
Test: None
Change-Id: I7588f5fc1249a688e95b5e8dfa078528e6451f63