Check for paths with spaces in ndk-build.cmd.
There's a check for this in build-local.mk, but it doesn't work on
Windows. See the comment for details.

The same check is also added to the auto-generated ndk-build.cmd
launcher in the root of the NDK, since if the NDK is installed to a path
with spaces it won't be able to invoke build/ndk-build.cmd successfully
to show this error.

Bug: https://github.com/android/ndk/issues/1400
Test: installed NDK to a path with spaces on windows, ran ndk-build
Change-Id: I6cd79231890b452e65d9a6923121155d623a7c71