Remove libc++ testing support.
This is now tested as part of the toolchain release. We don't have the
source of the tests that match the toolchain's libc++, so we can't run
them any more. Even if we could, it's not useful to run them this
late.

Plus, look at all the code that we get to delete! This was always a
gross hack and I'm glad to see it go. This also makes our test
artifacts significantly smaller, and building/running the tests it
*much* faster (7 minutes to build, 40 seconds to run).

Bug: None
Test: poetry run ./run_tests.py --rebuild.
Change-Id: I512417dd9d088c5b797bbc034a387670cffcf955