Convert ndk-gdb to a PEP-441 zip application.
Python supports executing zipped applications and running them in that
manner makes dependency discovery easier, since the application
directory will automatically be added to sys.path.

We do have to stop using hyphens in our script names though, because
those aren't valid python modules.

This does potentially allow us to do more cool things. If we moved
ndkgdb.py into ndk.tools, we could include the rest of the ndk package
so we can use things like ndk.hosts and ndk.abis. Moving it there also
makes it easier to ensure that it'll be covered by our linters and
tests, which was the real motivation for this change.

Bug: None
Test: ./checkbuild.py ndk-gdb ndk-gdb-shortcut && $NDK/ndk-gdb --help
Change-Id: Ibdd0c2d42a2eccaf46d188221421601f740b02d9