Move ndk-stack tests to a more testable location.
The hyphen in the directory name made it an invalid python package path
which meant these tests could only be found if the full path was given.

We ought to be putting all our python tests in the same directory anyway
so we don't have to update the test path in checkbuild.py every time we
add a new test.

This may not be a permanent location. This project is still a really
funky layout for something that's at least pretending to be a python
application (and vscode is baffled by the import of what it thinks is
the same module), but it may be worthwhile to split some of the python
code out so it can be shared with other projects anyway. This will do
for now.

Bug: None
Test: pytest tests/pytest
Change-Id: Ib12e1d6cf509291642806039ffd7b32e5aa35ccb