Improve accuracy of the cmake exports test.
Previously this would check that the values were defined, but wouldn't
fail if they were defined incorrectly. This checks that the programs
come from the NDK and not from somewhere else (see the bug).

Unfortunately this test still would not have prevent the bug because we
do not (and cannot feasibly) test with every version of CMake, and the
broken behavior reported in the bug is not present with CMake 3.22
(which we test with) but does in 3.26 (which we don't have available for
testing).

Bug: https://github.com/android/ndk/issues/1698
Test: This is the test
Change-Id: I8e978d5f6935d392a6c2e836429c4e7ac2366c27