Remove a probably useless test.
The history has been GC'd (code.google.com bug URLs are no longer
forwarded to buganizer) so I can't tell, but as best as I can tell this
test is just checking that SL_IID_ANDROIDBUFFERQUEUESOURCE is defined.
This belongs in CTS, not here.

The reason I'm here is because the release branch build (windows test
build) isn't working and our best guess at why is because we've hit a
path length limit. This test has a very long name and seems to be the
only one reliably failing. I could shorten the test name instead, but
this test looks useless, so might as well remove it.

Bug: http://b/285349982
Test: None
Change-Id: I8d80987176a531db2889543aaef84fa461e7f1ae