Use unavailable symbol with API level 100 for test
For the weak_symbols tests, the use of the Android trace API as the weak
symbol to verify against only worked because the existing architectures
had a minimum SDK version of 21, lower than the version 29 the API was
introduced in.

To accomodate for future architectures like riscv64, the test is
modified to rely on a fake API 'AFoo' that is marked as introduced in
version 100, to allow for weak symbol validation into the future.

Test: checkbuild.py
Change-Id: I52b918282178804db4dd19f40ec344e474ccc015