Fix ndk-stack, remove test reliance on mocks.
ndk-stack was moved to being a zipapp, which changed the path of
__file__ relative to the root of the NDK.

This was covered by a test (97% line coverage, even), but the test was
useless because it was only testing mocks, and the mocks weren't
updated.

The "system test" has been updated to test the real ndk-stack entry
point in the NDK rather than importing the source and mocking dependency
discovery. This unfortunately means that running the tests now depends
on more or less the whole NDK, but that's true of everything else in the
tests/ directory anyway. If that's ever a problem, we should split up
the tests so that the test of the NDK itself and the tests of the ndk
python package are kept separately.

Bug: https://github.com/android/ndk/issues/1751
Test: fixed the test, also tested manually
Change-Id: I77076d1c3fc8ac5a05a93dbe1ce3488874283143