Correct the r26 LLVM changelog.
We included a few changelog updates for the "d" toolchain, and even
included the new prebuilts, but neglected to actually update the NDK's
config.

The toolchain itself needs to be updated directly in the release branch
because master has already moved on for rv64 support.
https://r.android.com/2775899 does the update (but will be abandoned
once coastguard picks it up).

Bug: https://github.com/android/ndk/issues/1907
Bug: https://github.com/android/ndk/issues/1909
Test: None
Change-Id: Ibff877a664290979aa8488341f92f145797889a8