Don't build rv64 by default.
It sounds like r27 is actually going to be a rolling release rather than
next year's LTS, so it'll be going out quite a bit sooner than we
expected. rv64 support is still to young to be on by default, so flip it
off, and add a note to the changelog warning users to not worry about it
just yet.

Bug: None
Test: None
Change-Id: Ie1de976074d6a5526a5027cde7c19ee496dc3b9e